requirements.txt
tqdm
pyzipper
Pillow